Zen_multichatbot/requirements.txt
python-telegram-bot>=20.0
discord.py>=2.0
tweepy>=4.0
PyYAML>=6.0
